Rename basic jobs, replace - with _
The multijob environment variable LAST_TRIGGERED_JOB_NAME has all the -
signs replaced by _. It is not possible to get to the real job name from
the variables. This makes it impossible to create correct URLs to the
console from the jobs.

The pragmatic solution is to rename the jobs to have an underscore in
their name instead of the dashes.

Signed-off-by: Niels de Vos <ndevos@redhat.com>